:feat document: suite correction

+++ app/Controllers/Document.php
@@ -19,27 +19,7 @@ class Document extends PpciController
     }
     function write($origin)
     {
-        $ok = false;
-        if ($_SESSION["userRights"]["param"] == 1) {
-            $ok = true;
-        }
-        /*collection*/
-        if (
-            !$ok &&
-            $origin == "collection" && ($_SESSION["userRights"]["param"] == 1
-                || ($_SESSION["userRights"]["collection"] == 1 && !empty($_SESSION["collections"][$_REQUEST["collection_id"]]))
-            )
-        ) {
-            $ok = true;
-        }
-        /* sample */
-        if (!$ok && $origin == "sample" && !empty($_SESSION["collections"][$_REQUEST["collection_id"]])) {
-            $ok = true;
-        }
-        if (!$ok && $origin != "collection" && $origin != "sample") {
-            $ok = true;
-        }
-        if (!$ok) {
+        if (!$this->verifyRights($origin)) {
             $this->message->set(_("Vous ne disposez pas des droits suffisants pour cette opération"), true);
             return $this->returnToOrigin($origin, true);
         } else {
@@ -48,7 +28,12 @@ class Document extends PpciController
     }
     function delete($origin)
     {
-        return $this->returnToOrigin($origin,  $this->lib->delete());
+        if (!$this->verifyRights($origin)) {
+            $this->message->set(_("Vous ne disposez pas des droits suffisants pour cette opération"), true);
+            return $this->returnToOrigin($origin, true);
+        } else {
+            return $this->returnToOrigin($origin,  $this->lib->delete());
+        }
     }
     function get()
     {
@@ -96,6 +81,18 @@ class Document extends PpciController
             } else {
                 return $lib->display();
             }
+        }
+    }
+    function verifyRights($origin)
+    {
+        if ($origin == "sample") {
+            $lib = new Sample;
+            return $lib->verifyRights($_REQUEST["uid"]);
+        } elseif ($origin == "collection") {
+            $lib = new Collection;
+            return $lib->verifyRights($_REQUEST["collection_id"]);
+        } else {
+            return true;
         }
     }
 }
